backend/requirements.txt
fastapi[standard] >= 0.115.6
requests >= 2.32.3
python-dotenv >= 1.0.1
sqlalchemy >= 2.0.36
cryptography >= 44.0.0
bcrypt >= 4.2.1
pyotp >= 2.9.0
pyjwt >= 2.10.1
requests >= 2.32.3
redis >= 5.2.1
prometheus-client >= 0.21.1
pandas >= 2.2.3

tqdm >= 4.67.1
scikit-learn >= 1.7.0
statsmodels >= 0.14.4
xgboost >= 3.0.2
tensorflow >= 2.19.0
darts >= 0.35.0

prophet >= 1.1.7
pmdarima >= 2.0.4
